+++ test/eca/features/completion_test.clj
@@ -1,7 +1,8 @@
 (ns eca.features.completion-test
   (:require
    [clojure.test :refer [deftest is testing]]
-   [eca.features.completion :as f.completion]))
+   [eca.features.completion :as f.completion]
+   [eca.test-helper :as h]))
 
 (deftest normalize-code-result-test
   (testing "triple backticks with language label"
@@ -23,3 +24,19 @@
   (testing "preserve leading spaces when no fences"
     (is (= "    (println \"indented\")"
            (#'f.completion/normalize-code-result "    (println \"indented\")")))))
+
+(deftest complete-with-nil-model-test
+  (testing "Returns clean error when no completion model is configured (nil full-model)"
+    (h/reset-components!)
+    (h/config! {:env "test"})
+    (let [result (f.completion/complete
+                  {:doc-text "(defn foo [] 1)"
+                   :doc-version 1
+                   :position {:line 1 :character 16}}
+                  (h/db*)
+                  (h/config)
+                  (h/messenger)
+                  (h/metrics))]
+      (is (some? (:error result)) "Should return an error map")
+      (is (= :warning (get-in result [:error :type])))
+      (is (string? (get-in result [:error :message]))))))